refactor: update file structure and improve header navigation
