+++ src/pages/plats_page.jsx
